requirements.txt
mysql-connector-python>=8.0.0
pandas>=2.0.0
matplotlib>=3.7.0

